RavenDB-25233 : AiUsage token counts should use long instead of int to prevent integer overflow

+++ src/Raven.Client/Documents/Operations/AI/AiUsage.cs
@@ -6,17 +6,17 @@ namespace Raven.Client.Documents.Operations.AI;
 
 public class AiUsage : IDynamicJsonValueConvertible
 {
-    public int PromptTokens { get; set; }
-    public int CompletionTokens { get; set; }
-    public int TotalTokens { get; set; }
-    public int CachedTokens { get; set; }
-    public int ReasoningTokens { get; set; }
+    public long PromptTokens { get; set; }
+    public long CompletionTokens { get; set; }
+    public long TotalTokens { get; set; }
+    public long CachedTokens { get; set; }
+    public long ReasoningTokens { get; set; }
 
     internal void UpdateFrom(BlittableJsonReaderObject json)
     {
-        json.TryGet("prompt_tokens", out int promptTokens);
-        json.TryGet("completion_tokens", out int completionTokens);
-        json.TryGet("total_tokens", out int totalTokens);
+        json.TryGet("prompt_tokens", out long promptTokens);
+        json.TryGet("completion_tokens", out long completionTokens);
+        json.TryGet("total_tokens", out long totalTokens);
 
         PromptTokens += promptTokens;
         CompletionTokens += completionTokens;
@@ -24,13 +24,13 @@ public class AiUsage : IDynamicJsonValueConvertible
 
         if (json.TryGet("prompt_tokens_details", out BlittableJsonReaderObject promptDetails) && promptDetails != null)
         {
-            if (promptDetails.TryGet("cached_tokens", out int cachedTokens))
+            if (promptDetails.TryGet("cached_tokens", out long cachedTokens))
                 CachedTokens += cachedTokens;
         }
 
         if (json.TryGet("completion_tokens_details", out BlittableJsonReaderObject completionTokensDetails) && completionTokensDetails != null)
         {
-            if (completionTokensDetails.TryGet("reasoning_tokens", out int reasoningTokens))
+            if (completionTokensDetails.TryGet("reasoning_tokens", out long reasoningTokens))
             {
                 ReasoningTokens += reasoningTokens;
             }
